api/pyproject.toml
[tool.poetry]
name = "passport-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
django = "==4.2.6"
didkit = "*"
uritemplate = "*"
django-environ = "*"
django-ninja = "*"
django-ninja-jwt = "*"
django-ninja-extra = "*"
web3 = "*"
psycopg2-binary = "*"
djangorestframework-api-key = "==2.*"
requests = "*"
dag-cbor = "*"
django-ratelimit = "*"
ninja_schema = "*"
structlog = "*"
django-structlog = "*"
django-filter = "*"
boto3 = "*"
python-jose = "*"
uvarint = "*"
django-ace = "*"
async-timeout = "*"
aiohttp = "*"
redis = "^5.1.1"
siwe = "^4.4.0"
multiformats = "^0.3.1.post4"
django-storages = "^1.14.4"


[tool.poetry.group.dev.dependencies]
black = "*"
pre-commit = "*"
pylint = "*"
pylint-django = "*"
coverage = "*"
pytest-django = "*"
pytest-bdd = "*"
pytest-mock = "*"
django-stubs = "*"
pandas = "*"
faker = "*"
freezegun = "*"
pytest-asyncio = "^0.25.2"
ruff = "^0.11.2"


[tool.poetry.group.server.dependencies]
uvicorn = "^0.32.1"
django-extensions = "^3.2.3"
django-cors-headers = "^4.6.0"
social-auth-app-django = "^5.4.2"
pyarrow = "^18.0.0"
tqdm = "^4.67.0"
django-ipware = "^7.0.1"
whitenoise = "^6.8.2"
gunicorn = "^23.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
fixable = ["ALL"]
select = [
    # isort
    "I",
]
